src/requirements.txt
certifi==2023.7.22
charset-normalizer==3.3.1
idna==3.4
numpy==1.26.1
opencv-python==4.8.1.78
Pillow==10.1.0
requests==2.31.0
urllib3==2.0.7
